ai-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-multipart==0.0.9
pdfminer.six==20231228
python-docx==1.1.0
spacy==3.7.5
scikit-learn==1.5.1
sentence-transformers==3.0.1
numpy==1.26.4
pandas==2.2.2
pydantic==2.8.0
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
alembic==1.13.2
pytest==8.3.2
gunicorn==22.0.0
